added the cheakout proceedure

+++ Api/Api/Entities/Bid.cs
@@ -1,4 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿    using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Api.Entities
@@ -12,7 +12,7 @@ namespace Api.Entities
 
        public DateTime BidDate { get; set; }= DateTime.Now;
 
-        public double BidPrice { get; set; }
+        public long BidPrice { get; set; }
 
         public int AuctionID {  get; set; }
